+++ src/index.ts
@@ -2,6 +2,7 @@ export * from './types';
 export * from './combine';
 export * from './create-assign';
 export * from './create-table';
+export * from './create-map';
 export * from './create-loader';
 export * from './create-loader-table';
 
@@ -11,19 +12,21 @@ export { default as createAction } from './create-action';
 export { default as createReducer } from './create-reducer';
 export { default as createApp } from './create-app';
 import createTable from './create-table';
+import createMap from './create-map';
 import createLoader from './create-loader';
 import createLoaderTable from './create-table';
 import createAssign from './create-assign';
 export {
   createSlice,
   createTable,
+  createMap,
   createLoader,
   createLoaderTable,
   createAssign,
 };
 
 // All of these variable names are DEPRECATED
-const mapSlice = createTable;
+const mapSlice = createMap;
 const loadingSlice = createLoader;
 const loadingMapSlice = createLoaderTable;
 const assignSlice = createAssign;