+++ type-tests/slice.ts
@@ -3,7 +3,7 @@ import createSlice from '../src/create-slice';
 // testing InputWithName and 2 params
 const addFour = createSlice<number, { add: number }>({
   name: 'something',
-  reducts: { add: (state, p: number) => state + p },
+  reducers: { add: (state, p: number) => state + p },
   initialState: 0,
 });
 // $ExpectType { add: (payload: number) => Action<number, string>; }
@@ -16,7 +16,7 @@ addFour.name;
 // testing InputWithName and no params
 const addFive = createSlice({
   name: 'something' as const,
-  reducts: { add: (state, p: number) => state + p },
+  reducers: { add: (state, p: number) => state + p },
   initialState: 0,
 });
 // $ExpectType { add: (payload: number) => Action<number, string>; }
@@ -30,7 +30,7 @@ addFive.name;
 let eightSlice = 'something';
 const addEight = createSlice({
   name: eightSlice,
-  reducts: { add: (state, p: number) => state + p },
+  reducers: { add: (state, p: number) => state + p },
   initialState: 0,
 });
 // $ExpectType { add: (payload: number) => Action<number, string>; }